first attempt of creating a custom hook for the useReducer hook of the todo-app
